refactor: rename to @cgr/project-config
