+++ apps/web/package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "devDependencies": {
-    "@cgr/config": "workspace:*",
+    "@cgr/project-config": "workspace:*",
     "@next/bundle-analyzer": "^13.1.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/react-hooks": "8.0.1",
