+++ packages/project-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@cgr/config",
+  "name": "@cgr/project-config",
   "version": "0.0.0",
   "private": true
 }
